Add add_flag.
WIP:add_options with string match and document.

+++ lib_test.cc
@@ -3,4 +3,4 @@
 
 #include "options_parser/options_parser_lib.h"
 
-OPTIONS_PARSER_FLAGS_DEFINE(int, flag_float, 0, "A flag with float value");
+OPTIONS_PARSER_FLAGS_DEFINE(float, flag_float, 0, "A flag with float value");
